[routes]organise fileUploader router
